repackage version model
